+++ modules/default/calendar/calendarutils.js
@@ -498,8 +498,7 @@ const CalendarUtils = {
 			return a.startDate - b.startDate;
 		});
 
-		let maxEvents = newEvents.slice(0, config.maximumEntries);
-		return maxEvents;
+		return newEvents;
 	},
 
 	/**